+++ src/modules/InvitationForPunchOut/views/SearchIPO/SearchIpo.style.tsx
